requirements.txt
# FastAPI and related packages
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3

# OpenAI API
openai==1.3.0

# Environment variables
python-dotenv==1.0.0

# HTTP client
httpx==0.25.1
